Skip loading Pods and Gems when checksum is not valid anymore.

+++ runner-vm-pre-build.sh
@@ -28,23 +28,27 @@ if [ -z $SKIP_RUNNER_CACHE ]; then
 	PODS_ARCHIVE="Pods.tar.bz2"
 	GEMS_ARCHIVE="Gems.tar.bz2"
 	CACHE_PATH="$MOUNT_PATH/${CI_RUNNER_ID}/${CI_PROJECT_NAME}"
+	
+	PODS_CHECKSUM_PATH="$CACHE_PATH/Podfile.checksum"
+	PODFILE_LOCK_CHECKSUM=`test -s Podfile.lock && md5 -q Podfile.lock`
 
-	if [ -e ${CACHE_PATH}/${PODS_ARCHIVE} ]; then
+	if [ -e ${CACHE_PATH}/${PODS_ARCHIVE} ] && [ "$PODFILE_LOCK_CHECKSUM" == "$(cat $PODS_CHECKSUM_PATH)" ]; then
 		echo "--- Loading pods cache..."
 		time cp ${CACHE_PATH}/${PODS_ARCHIVE} $PODS_ARCHIVE
 		time openssl enc -aes-256-cbc -k $AUTO_CLOSE_TOKEN -d -in $PODS_ARCHIVE | pbzip2 -dc | tar xf -
-
-		PODS_CHECKSUM_PATH="$CACHE_PATH/Podfile.checksum"
-		export CACHED_PODFILE_LOCK_CHECKSUM=`test -s $PODS_CHECKSUM_PATH && cat $PODS_CHECKSUM_PATH`
+	else
+	  echo "--- Skipping Pods cache. Podfile.lock checksum is not valid anymore."
 	fi
 
-	if [ -e ${CACHE_PATH}/${GEMS_ARCHIVE} ]; then
+	BUNDLE_CHECKSUM_PATH="$CACHE_PATH/Gemfile.checksum"
+	GEMFILE_LOCK_CHECKSUM=`test -s Gemfile.lock && md5 -q Gemfile.lock`
+	
+	if [ -e ${CACHE_PATH}/${GEMS_ARCHIVE} ] && [ "$GEMFILE_LOCK_CHECKSUM" == "$(cat $BUNDLE_CHECKSUM_PATH)" ]; then
 		echo "--- Loading gems cache..."
 		time cp ${CACHE_PATH}/${GEMS_ARCHIVE} $GEMS_ARCHIVE
 		time openssl enc -aes-256-cbc -k $AUTO_CLOSE_TOKEN -d -in $GEMS_ARCHIVE | pbzip2 -dc | tar xf -
-
-		BUNDLE_CHECKSUM_PATH="$CACHE_PATH/Gemfile.checksum"
-		export CACHED_GEMFILE_LOCK_CHECKSUM=`test -s $BUNDLE_CHECKSUM_PATH && cat $BUNDLE_CHECKSUM_PATH`
+	else
+	  echo "--- Skipping Gems cache. Gemfile.lock checksum is not valid anymore."
 	fi
 
 	echo "--- Downloading cache finished..."
